Fix operation within loop or w/ non-interactive use
This patch should permit display of conf.mat into the currently open graphics object (e.g. jpeg file) when executed in script or loop.

See https://stackoverflow.com/questions/7981507/lattice-plots-not-working-inside-a-function and https://stackoverflow.com/questions/26643852/ggplot-plots-in-scripts-do-not-display-in-rstudio

alternatively `return (gobj)` instead of `print (gobj)` so that users can manually modify the chart themselves prior to `print`ing the chart themselves.